refactor: share state validator
